tensorflow-serving-grpc/client/client_requirements.txt
numpy==1.22.4
tensorflow==2.11.0
tensorflow_serving_api==2.11.0
